# app/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenMark API",
    "description": "API for PDF document viewing with annotation capabilities. OpenMark provides secure access to PDF documents with support for notes, highlights, and collaborative annotations.\n\n## Authentication\n\nOpenMark uses two types of tokens:\n\n1. **Auth Token**: Standard JWT token obtained via `/api/authenticate`. Used for most API calls.\n2. **Document Access Token (DAT)**: Self-contained JWT for document access. Survives page refresh.\n\n## Quick Start\n\n### Option 1: Quick View (Single Call)\n```\nPOST /api/quickView\n{\n  \"username\": \"user\",\n  \"password\": \"pass\",\n  \"documentId\": \"sample\"\n}\n```\n\n### Option 2: Three-Step Process\n1. Authenticate → Get token\n2. Request document → Get DAT\n3. View document with DAT",
    "version": "1.0.0",
    "contact": {
      "name": "OpenMark Support"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server"
    }
  ],
...
  "paths": {
    "/authenticate": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Authenticate user",
        "description": "Authenticate a user with username and password. Returns a JWT token for subsequent API calls.",
        "operationId": "authenticate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              },
              "example": {
                "username": "admin",
                "password": "admin123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing username or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Logout user",
        "description": "Invalidate the current authentication token.",
        "operationId": "logout",
        "security": [
          {"bearerAuth": []}
        ],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/quickView": {
      "post": {
        "tags": ["Documents"],
        "summary": "Quick view document (single call)",
        "description": "Authenticate and request a document in one call. Returns a viewer URL with Document Access Token (DAT). This is the recommended method for simple integrations.\n\nThe returned URL can be used directly in an iframe or opened in a new tab. The DAT is self-contained and allows page refresh without losing access.",
        "operationId": "quickView",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickViewRequest"
              },
              "example": {
                "username": "admin",
                "password": "admin123",
                "documentId": "sample",
                "hideAnnotationsTools": false,
                "hideAnnotations": false,
                "hideLogo": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document ready for viewing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickViewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/requestDocument": {
      "post": {
        "tags": ["Documents"],
        "summary": "Request document access",
        "description": "Request access to a PDF document. Returns a temporary document ID and Document Access Token (DAT). Use this endpoint after authenticating to get document access.",
        "operationId": "requestDocument",
        "security": [
          {"bearerAuth": []}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestDocumentRequest"
              },
              "example": {
                "documentId": "sample",
                "hideAnnotationsTools": false,
                "hideAnnotations": false,
                "hideLogo": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document access granted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestDocumentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing documentId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/documentStatus/{tempDocId}": {
      "get": {
        "tags": ["Documents"],
        "summary": "Get document download status",
        "description": "Check the download/caching status of a requested document. Status can be: pending, downloading, ready, or error.",
        "operationId": "getDocumentStatus",
        "parameters": [
          {
            "name": "tempDocId",
            "in": "path",
            "required": true,
            "description": "Temporary document ID returned from requestDocument or quickView",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dat",
            "in": "query",
            "required": false,
            "description": "Document Access Token (alternative to Bearer token)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {"bearerAuth": []},
          {"datAuth": []}
        ],
        "responses": {
          "200": {
            "description": "Document status retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/viewDocument": {
      "get": {
        "tags": ["Documents"],
        "summary": "View document in PDF viewer",
        "description": "Redirect to the PDF viewer with the specified document. Use the Document Access Token (DAT) for authentication.",
        "operationId": "viewDocument",
        "parameters": [
          {
            "name": "dat",
            "in": "query",
            "required": true,
            "description": "Document Access Token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to PDF viewer"
          },
          "401": {
            "description": "Invalid or expired DAT"
          }
        }
      }
    },
    "/getAnnotations": {
      "get": {
        "tags": ["Annotations"],
        "summary": "Get document annotations",
        "description": "Retrieve all annotations (notes and highlights) for a specific document.",
        "operationId": "getAnnotations",
        "security": [
          {"bearerAuth": []},
          {"datAuth": []}
        ],
        "parameters": [
          {
            "name": "documentId",
            "in": "query",
            "required": true,
            "description": "Document ID to get annotations for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Annotations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAnnotationsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing documentId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/saveAnnotations": {
      "post": {
        "tags": ["Annotations"],
        "summary": "Save document annotations",
        "description": "Save annotations (notes and highlights) for a document. Replaces existing annotations for the user/document combination.",
        "operationId": "saveAnnotations",
        "security": [
          {"bearerAuth": []},
          {"datAuth": []}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAnnotationsRequest"
              },
              "example": {
                "documentId": "sample",
                "annotations": {
                  "notes": [
                    {
                      "id": "note-1",
                      "page": 1,
                      "x": 100,
                      "y": 200,
                      "content": "Important section",
                      "color": "#ffeb3b"
                    }
                  ],
                  "highlights": [
                    {
                      "id": "highlight-1",
                      "page": 1,
                      "rects": [{"x": 50, "y": 100, "width": 200, "height": 20}],
                      "color": "#ffff00"
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Annotations saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Annotations saved successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to save annotations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/statistics": {
      "get": {
        "tags": ["User Data"],
        "summary": "Get user statistics",
        "description": "Retrieve usage statistics for the authenticated user including documents viewed, notes created, and highlights created.",
        "operationId": "getStatistics",
        "security": [
          {"bearerAuth": []}
        ],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /authenticate endpoint"
      },
      "datAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "dat",
        "description": "Document Access Token for document-specific operations"
      }
    },
    "schemas": {
      "AuthRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "User's username"
          },
          "password": {
            "type": "string",
            "description": "User's password"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "token": {
            "type": "string",
            "description": "JWT authentication token"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "Token expiration timestamp"
          }
        }
      },
...
    }
  }
}